Separate methods into modules
