+++ src/index.js
@@ -1,3 +1,6 @@
+import { dom, convertTemp } from './domModule';
+import { cap, convertNull, decimals } from './conversionModule';
+
 const request = JSON.parse(localStorage.getItem('request')) || [];
 
 const getUrl = (getCity, getState) => {
@@ -14,18 +17,6 @@ const getUrl = (getCity, getState) => {
     return `${root}${city}${key}`
   } else {
     return `${root}${city},${state}${key}`
-  }
-};
-
-const cap = (string) => {
-  return string.charAt(0).toUpperCase() + string.slice(1);
-}
-
-const convertNull = (string) => {
-  if(string === null){
-    return '';
-  } else {
-    return string;
   }
 };
 
@@ -55,14 +46,6 @@ const errorHanlder = (err) => {
   );
 };
 
-const decimals = (number) => {
-  if(number % 1 !== 0) {
-    return parseFloat(number.toFixed(2));
-  } else {
-    return number
-  }
-};
-
 const dataHandler = (data, city, state, pref) => {
   if (data.cod === 200) {
     const weather = newWeather(data, city, state, pref);
@@ -86,124 +69,5 @@ const newWeather = (weatherData, city, state, pref) => {
   return { city, state, kelvin, celcius, fahrenheit, pref, description, icon };
 };
 
-const placeValues = (city, state) => {
-  if(state.length === 0) {
-    return `The weather in ${city} is:`
-  } else if (city.length === 0) {
-    return `The weather in ${cap(state)} is:`
-  } else {
-    return location.textContent = `The weather in ${city}, ${cap(state)} is:`
-  }
-};
-
-const showTemp = (data) => {
-  switch (data.pref) {
-    case '1':
-      return `${data.celcius} °C`
-    case '2':
-      return `${data.fahrenheit} °F`
-    case '3':
-      return `${data.kelvin} °K`
-  }
-};
-
-const dataTypeSet = (type, compare) => {
-  if(type === compare) {
-    return 'temp-switch';
-  } else {
-    return 'temp-switch temp-click';
-  }
-};
-
-const dom = (weatherData) => {
-  const pageContent = document.getElementById('page-content');
-  const fragment = document.createDocumentFragment();
-  const location = document.createElement('h1');
-  location.textContent = placeValues(weatherData.city, weatherData.state);
-
-
-  const temp = document.createElement('span');
-  temp.id = 'temp';
-  temp.textContent = showTemp(weatherData);
-
-  const switchDiv = document.createElement('div');
-  switchDiv.id = 'switch-type';
-  switchDiv.classList = 'switch';
-
-  const description = document.createElement('p');
-  description.textContent = weatherData.description;
-  const icon = document.createElement('img');
-  icon.src = weatherData.icon;
-
-  const c = document.createElement('span');
-  c.textContent = '°C';
-  c.id = 'celcius';
-  c.classList = dataTypeSet('1', weatherData.pref);
-  c.setAttribute('data-type', 'celcius');
-
-  const f = document.createElement('span');
-  f.textContent = '°F';
-  f.id = 'fahrenheit';
-  f.classList = dataTypeSet('2', weatherData.pref);
-  f.classList.add('middle-temp');
-  f.setAttribute('data-type', 'fahrenheit');
-
-  const k = document.createElement('span');
-  k.textContent = '°K';
-  k.id = 'kelvin';
-  k.setAttribute('data-type', 'kelvin');
-  k.classList = dataTypeSet('3', weatherData.pref);
-
-  fragment.appendChild(location);
-  fragment.appendChild(description);
-  fragment.appendChild(icon);
-  fragment.appendChild(temp);
-  switchDiv.appendChild(c);
-  switchDiv.appendChild(f);
-  switchDiv.appendChild(k);
-  fragment.appendChild(switchDiv);
-  pageContent.appendChild(fragment);
-};
-
-const showWeather = (data) => {
-  if (data.length > 0) {
-    dom(data[0]);
-  }
-};
-
-showWeather(request);
-
-const convertTemp = (data) => {
-  const tempsSwitch = document.querySelectorAll('.temp-switch');
-  tempsSwitch.forEach(temps => {
-    temps.addEventListener('click', () => {
-      const tempTypeData = temps.dataset.type;
-      const temp = document.getElementById('temp');
-      switch(tempTypeData) {
-        case 'celcius':
-          temp.textContent = `${data.celcius} C°`;
-          resetClasses(tempsSwitch);
-          temps.classList.remove('temp-click');
-          break;
-        case 'kelvin':
-          temp.textContent = `${data.kelvin} K°`;
-          resetClasses(tempsSwitch);
-          temps.classList.remove('temp-click');
-          break;
-        case 'fahrenheit':
-          temp.textContent = `${data.fahrenheit} F°`;
-          resetClasses(tempsSwitch);
-          temps.classList.remove('temp-click');
-          break;
-      }
-    });
-  });
-};
-
-const resetClasses = (tempsSwitch) => {
-  tempsSwitch.forEach(temps => {
-    temps.classList.add('temp-click');
-  });
-};
-
+dom(request[0]);
 convertTemp(request[0]);
